Remove standalone experimental package
