+++ .changeset/byos-local-workbench.md
@@ -1,8 +1,7 @@
 ---
 '@composio/core': patch
-'@composio/experimental': patch
 ---
 
-Add experimental BYO local workbench primitives under `@composio/core/experimental`, plus an optional `@composio/experimental` package with an E2B sandbox adapter.
+Add experimental BYO local workbench primitives under `@composio/core/experimental`.
 
 The v0 local workbench authenticates sandbox tool execution with the developer's Composio project API key, keeps sandbox provider selection SDK-local, and exposes a provider-agnostic `SandboxProvider` contract for additional adapters.
