+++ pnpm-workspace.yaml
@@ -10,7 +10,6 @@ packages:
   - ts/packages/cli-keyring
   - ts/packages/cli-local-tools
   - ts/packages/core
-  - ts/packages/experimental
   - ts/packages/json-schema-to-zod
   - ts/packages/providers/*
   - ts/packages/ts-builders
@@ -32,7 +31,6 @@ catalog:
   '@effect/platform-node': ^0.107.0
   '@effect/platform-node-shared': ^0.60.0
   '@effect/vitest': ^0.29.0
-  '@e2b/code-interpreter': ^2.6.1
   '@mastra/core': ^1.42.0
   '@modelcontextprotocol/sdk': ^1.29.0
   '@types/bun': ^1.3.14
