--- ts/packages/experimental/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@composio/experimental",
-  "version": "0.0.1",
-  "description": "Experimental Composio SDK capabilities.",
-  "main": "src/index.ts",
-  "type": "module",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ComposioHQ/composio.git",
-    "directory": "ts/packages/experimental"
-  },
-  "bugs": {
-    "url": "https://github.com/ComposioHQ/composio/issues"
-  },
-  "homepage": "https://github.com/ComposioHQ/composio/tree/main/ts/packages/experimental#readme",
-  "publishConfig": {
-    "access": "public",
-    "main": "dist/index.mjs",
-    "types": "dist/index.d.mts"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.mts",
-      "default": "./dist/index.mjs"
-    }
-  },
-  "files": [
-    "README.md",
-    "dist"
-  ],
-  "scripts": {
-    "clean": "git clean -xdf node_modules",
-    "build": "pnpm exec tsdown",
-    "test": "vitest run",
-    "typecheck": "pnpm exec tsgo --noEmit --skipLibCheck"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "peerDependencies": {
-    "@composio/core": ">=0.11.0 <1.0.0",
-    "@e2b/code-interpreter": "catalog:"
-  },
-  "devDependencies": {
-    "@composio/core": "workspace:*",
-    "@e2b/code-interpreter": "catalog:",
-    "tsdown": "catalog:",
-    "typescript": "catalog:",
-    "vitest": "catalog:"
-  }
-}
