--- ts/packages/experimental/src/index.ts
@@ -1,19 +1,0 @@
-export {
-  COMPOSIO_WORKBENCH_HELPER_PATH,
-  experimental_createLocalWorkbenchSession,
-  experimental_createWorkbenchEnv,
-  experimental_createWorkbenchHelperSource,
-} from '@composio/core/experimental';
-export type {
-  LocalWorkbenchConfig,
-  LocalWorkbenchSession,
-  SandboxExecutionOptions,
-  SandboxExecutionResult,
-  SandboxProvider,
-  SandboxProviderName,
-  SandboxProvisionContext,
-  WorkbenchEnvOptions,
-  WorkbenchHelperSourceOptions,
-} from '@composio/core/experimental';
-export { experimental_e2bSandbox } from './e2b';
-export type { E2BSandboxOptions } from './e2b';
