--- ts/packages/experimental/src/local-workbench.ts
@@ -1,1 +1,0 @@
-export { experimental_createLocalWorkbenchSession } from '@composio/core/experimental';
