--- ts/packages/experimental/src/shim.ts
@@ -1,9 +1,0 @@
-export {
-  COMPOSIO_WORKBENCH_HELPER_PATH,
-  experimental_createWorkbenchEnv,
-  experimental_createWorkbenchHelperSource,
-} from '@composio/core/experimental';
-export type {
-  WorkbenchEnvOptions,
-  WorkbenchHelperSourceOptions,
-} from '@composio/core/experimental';
