--- ts/packages/experimental/src/types.ts
@@ -1,9 +1,0 @@
-export type {
-  LocalWorkbenchConfig,
-  LocalWorkbenchSession,
-  SandboxExecutionOptions,
-  SandboxExecutionResult,
-  SandboxProvider,
-  SandboxProviderName,
-  SandboxProvisionContext,
-} from '@composio/core/experimental';
