--- ts/packages/experimental/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "target": "es2022",
-    "module": "esnext",
-    "strict": true,
-    "esModuleInterop": true,
-    "moduleResolution": "bundler",
-    "skipLibCheck": true
-  },
-  "include": ["src", "test"]
-}
