--- ts/packages/experimental/tsdown.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'tsdown';
-import { baseConfig } from '../../../tsdown.config.base.ts';
-
-export default defineConfig({
-  ...baseConfig,
-  tsconfig: 'tsconfig.json',
-});
